backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
pydantic==2.7.4
httpx==0.27.0
python-dotenv==1.0.1

psycopg[binary]==3.2.1
psycopg_pool==3.2.1
PyJWT==2.8.0